% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeImages.R
\name{analyzeImages}
\alias{analyzeImages}
\title{Analyze Multiple Images}
\usage{
analyzeImages(
  path,
  paper_dim = c(76, 26),
  distance = 0.7,
  display.it = FALSE,
  cl = NULL
)
}
\arguments{
\item{path}{A character giving the path to the folder containing the
images (only!) of the papers.}

\item{paper_dim}{A numeric vector of length 2 indicating the length and
width (in mm) of the water-sensitive paper.}

\item{distance}{A numeric value indicating the distance from the nozzle
to the target, in order to estimate the spray drift (\%) using the
'German model' for field crops.}

\item{display.it}{(Logical) Should the image with the bounding box of the
paper be displayed? Default is \code{TRUE}.}

\item{cl}{An integer indicating the number of parallel processes or an
object created by \code{\link[parallel:makeCluster]{parallel::makeCluster()}}. Default if \code{NULL}.}
}
\value{
A list of objects of class \code{hydropaper}. See more in
\code{\link[=analyzePaper]{analyzePaper()}}.
}
\description{
Detect the water-sensitive paper from multiple images and do all the
analyses in a run. Parallel processing is allowed.
}
\examples{
path <- system.file('images', package = 'hydropaper')
list.files(path)
analyzeImages(path)

}
