% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzePaper.R
\name{analyzePaper}
\alias{analyzePaper}
\title{Analyze the Water-Sensitive Paper}
\usage{
analyzePaper(x, paper_dim = c(76, 26), distance = 0.7, display.it = TRUE)
}
\arguments{
\item{x}{Either an object of class \code{Image} from package \link{EBImage}
or a character indicating the path to the image file (.jpg, .jpeg, .png, .tif).}

\item{paper_dim}{A numeric vector of length 2 indicating the length and
width (in mm) of the water-sensitive paper.}

\item{distance}{A numeric value indicating the distance from the nozzle(s)
to the target, in order to estimate the spray drift (\%) using the
'German model' for field crops.}

\item{display.it}{(Logical) Should the image with the bounding box of the
paper be displayed? Default is \code{TRUE}.}
}
\value{
Am object of class \code{hydropaper}, consisting of a \code{list}
of the following:
\describe{
\item{ndrops}{ An integer indicating the number of droplets found.}
\item{coverage}{ The percentage of wet area of the paper.}
\item{density}{ The number of droplets per squared centimeter.}
\item{vol}{ The volume of product deposited in the paper.
If \code{paper_dim}is given in milimeters, then \code{vol} is given in
microliters.}
\item{vol_ha}{ The extrapolation of volume deposited for one hectare.
If \code{paper_dim} is given in milimeters, then \code{vol_ha} is given
in L/ha.}
\item{md}{ The mean diameter of droplets. If \code{paper_dim} is given
in milimeters, then \code{md} is given in micrometers.}
\item{nmd}{ The numeric median diameter of droplets.
If \code{paper_dim} is given in milimeters, then \code{nmd} is given in
micrometers.}
\item{vmd}{ The volumetric median diameter of droplets.
If \code{paper_dim} is given in milimeters, then \code{vmd} is given in
micrometers.}
\item{d1}{ The 10th percentile of the volumetric median diameter of droplets.
If \code{paper_dim} is given in milimeters, then \code{d1} is given in
micrometers.}
\item{d9}{ The 90th percentile of the volumetric median diameter of droplets.
If \code{paper_dim} is given in milimeters, then \code{d9} is given in
micrometers.}
\item{RA}{ The relative amplitude: \code{(d9 - d1)/vmd}.}
\item{maxdrop}{ The maximum diameter of droplets.
If \code{paper_dim} is given in milimeters, then \code{maxdrop} is given in
micrometers.}
\item{mindrop}{ The minimum diameter of droplets.
If \code{paper_dim} is given in milimeters, then \code{mindrop} is given in
micrometers.}
\item{size_class}{ The percentage of droplets in each of the following
three classes of diameter: <200, 200-400, >400 micrometers.}
\item{CVa}{ The coeficient of variation (\%) of droplet areas.}
\item{CVd}{ The coeficient of variation (\%) of droplet diameters.}
\item{r}{ A character giving a naive recomendation of which products
are suitable to be sprayed, based on \code{density}.}
\item{drift}{ A prediction of drift (\%) based on the 'German model' for
field crops (Rautmann et al., 2001), as function of \code{distance}.
This quantifies a 90th percentile of the drift values.}
\item{areas}{ A numeric vector containing area of each droplet found.
If \code{paper_dim} is given in milimeters, then \code{areas} is given in
squared milimeters.}
\item{diams}{ A numeric vector containing the diameter of each droplet
found. If \code{paper_dim} is given in milimeters, then \code{diams}
is given in squared micrometers.}
\item{binary}{ The binary mask of the segmentation, a numeric matrix.}
}
}
\description{
Evaluate the spray quality by analyzing an RGB-image of the water-sensitive
paper.
}
\examples{
path <- system.file('images', 'field1.jpg', package = 'hydropaper')
p1 <- detectPaper(path)

a <- analyzePaper(p1)
print(a)
plot(a)

}
\references{
Rautmann, D., M. Streloke, R. Winkler (2001) New basic drift values in
the authorization procedure for plant protection products.
In: R. Forster & M. Streloke, Workshop on Risk Assessment and Risk
Mitigation measures in the context of the authorization of plant
protection products (WORMM) 27.-29. September 1999. Mitteilungen
aus der Biologischen Bundesanstalt für Land- und Forstwirtschaft,
Berlin-Dahlem, Heft 381. 2001. 133-141.
}
\seealso{
\code{\link[=detectPaper]{detectPaper()}}
}
