% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipPaper.R
\name{clipPaper}
\alias{clipPaper}
\title{Clip Image to Extract the Water-Sensitive Paper}
\usage{
clipPaper(x)
}
\arguments{
\item{x}{Either an object of class \code{Image} from package \link{EBImage}
or a character indicating the path to the image file (.jpg, .jpeg, .png, .tif).}
}
\value{
An object of class \code{Image}.
}
\description{
Click on three points on the image to clip the water-sensitive paper,
for further analysis.
}
\examples{
path <- system.file('images', 'field1.jpg', package = 'hydropaper')
p1 <- clipPaper(path)
display(p1, "raster")

}
\seealso{
\code{\link[=detectPaper]{detectPaper()}}
}
