% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPaper.R
\name{detectPaper}
\alias{detectPaper}
\title{Detect the Water-Sensitive Paper}
\usage{
detectPaper(x, paper_dim = c(76, 26), display.it = TRUE)
}
\arguments{
\item{x}{Either an object of class \code{Image} from package \link{EBImage}
or a character indicating the path to the image file (.jpg, .jpeg, .png, .tif).}

\item{paper_dim}{A numeric vector of length 2 indicating the length and
width (in mm) of the water-sensitive paper.}

\item{display.it}{(Logical) Should the image with the bounding box of the
paper be displayed? Default is \code{TRUE}.}
}
\value{
An object of class \code{Image}. An additional attribute containing
the bounding box for the paper is stored. See Examples.
}
\description{
Detect the water-sensitive paper in RGB images using a pre-trained SURF
(Speed Up Robust Features) descriptor.
}
\examples{
path <- system.file('images', 'field1.jpg', package = 'hydropaper')
p1 <- detectPaper(path)
attr(p1, 'paper_bbox')


}
\seealso{

}
