\name{hydropaper-package}
\alias{hydropaper-package}
\alias{hydropaper}
\docType{package}

\title{ Water-Sensitive Paper Analysis }

\description{
Detect water-sensitive paper from field images and make
    analyses of spray performance.
}

\details{
\tabular{ll}{
Package: \tab hydropaper\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2021-11-29\cr
License: \tab GPL (>= 2)\cr
}
}

\note{
\emph{hydropaper} is an ongoing project.
Any and all criticism, comments and suggestions are welcome.
URL \url{https://arsilva87.github.io/hydropaper/}
}

\author{
Anderson Rodrigo da Silva, Wilhan Valasco dos Santos

Maintainer: Anderson Rodrigo da Silva <anderson.agro@hotmail.com>
}

\keyword{ package }
