% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{makeReport}
\alias{makeReport}
\title{Report Maker}
\usage{
makeReport(x)
}
\arguments{
\item{x}{An object of class \code{hydropaper}, created with the function
\code{\link[=analyzePaper]{analyzePaper()}}.}
}
\value{
A HTML (hydropaper_report.html) will be generated and stored the
working directory.
}
\description{
Create a document to report the results of the analyses of one or
more water-sensitive paper.
}
\examples{
path <- system.file('images', 'field1.jpg', package = 'hydropaper')
p1 <- detectPaper(path)
a <- analyzePaper(p1)
makeReport(a)  # check your working directory

}
